% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eerder_gekocht}
\alias{eerder_gekocht}
\title{Previously Bought}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 379 rows and 3 columns.
}
\usage{
eerder_gekocht
}
\description{
Data set with previously bought products
}
\keyword{datasets}
