% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_product_details.R
\name{get_product_details}
\alias{get_product_details}
\alias{get_product_url}
\alias{get_product_name}
\alias{get_product_unit}
\alias{get_product_name_unit}
\alias{get_product_image}
\title{Retrieve Product Attributes from Local Data}
\usage{
get_product_url(x)

get_product_name(x)

get_product_unit(x)

get_product_name_unit(x)

get_product_image(x)
}
\arguments{
\item{x}{Name or URL of a product.}
}
\description{
Provides offline access to product information such as name, unit, image, and URL that is available after running \code{\link[=update_product_list_from_html]{update_product_list_from_html()}}.
}
\section{Disclaimer}{



This open-source R package, \emph{shinyplus}, is an independent software project and is not affiliated with, endorsed by, or associated in any way with PLUS Retail B.V. or the website \href{https://www.plus.nl}{plus.nl}.

The package does \strong{not} contain or store any data from plus.nl. It contains \strong{no product information, pricing data, or content} originating from PLUS Retail B.V. or its systems. Users are required to provide their own data (e.g., product lists) manually.

All interactions with plus.nl occur through \strong{local, client-side browser automation}, using the \href{https://rstudio.github.io/chromote/}{\code{chromote}} package. This mechanism emulates a user operating a private browser session on their own machine. For example, when a user submits a basket to their PLUS shopping cart, this package simply opens a browser in the background and automates user-visible steps — such as logging in, navigating to product pages, and clicking buttons — exactly as a person would. No API calls or server-side data access is used or implied.

Any logos, icons, or brand references (e.g., the PLUS logo) used in the user interface are publicly available media assets sourced from \href{https://nl.wikipedia.org/wiki/PLUS_(Nederlandse_supermarkt)}{Wikipedia} under appropriate free-use licensing. All trademarks, service marks, and product names are the property of their respective owners.
}

