% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recently_bought}
\alias{recently_bought}
\title{Previously Bought}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 378 rows and 4 columns.
}
\usage{
recently_bought
}
\description{
Data set with previously bought products
}
\examples{
recently_bought
}
\keyword{datasets}
