% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email_weekmenu}
\alias{send_email_weekmenu}
\alias{send_email_basket}
\title{Send Through GMail}
\usage{
send_email_weekmenu(weekmenu, credentials = getOption("plus_credentials"))

send_email_basket(basket, credentials = getOption("plus_credentials"))
}
\arguments{
\item{weekmenu}{The \link{data.frame} of weekmenu to send}

\item{credentials}{Path to a YAML file containing fields \code{email} and/or \code{email_cc} (to send mail to), \code{gmail_user_name} (to send mail from) and \code{gmail_app_password}, or a \link{list} contains those names. Can be set with \code{options(plus_credentials = "...")}}

\item{basket}{The \link{data.frame} of basket items to send}
}
\description{
Send Through GMail
}
