% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyplus-package.R
\docType{package}
\name{shinyplus-package}
\alias{shinyplus-package}
\title{shinyplus: Shiny-empowered Automatic Shopping at PLUS}
\description{
Automates grocery shopping at PLUS
}
\section{Disclaimer}{

This open-source R package, \emph{shinyplus}, is independently developed and has no affiliation with, endorsement by, or association with PLUS Retail B.V. or the website \href{https://www.plus.nl}{plus.nl}.

The package merely uses \emph{client-side browser automation} via the \href{https://rstudio.github.io/chromote/}{\code{chromote}} package to interact with publicly available content on plus.nl, simulating user behaviour (such as logging in, browsing, and selecting products) in a private, ephemeral local browser session. This package does not contain any product info or data from PLUS. Any icons or images used (e.g. the PLUS logo) are sourced from \href{https://nl.wikipedia.org/wiki/PLUS_(Nederlandse_supermarkt)}{Wikipedia} under appropriate free-use licensing. All trademarks, product names, and brand assets remain the property of their respective owners.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/weberends/shinyplus}
  \item Report bugs at \url{https://github.com/weberends/shinyplus/issues}
}

}
\author{
\strong{Maintainer}: Wikje E. Berends-Hoekstra \email{w.e.berends-hoekstra@umcg.nl} (\href{https://orcid.org/0000-0001-7620-1800}{ORCID})

Authors:
\itemize{
  \item Matthijs S. Berends (\href{https://orcid.org/0000-0001-7620-1800}{ORCID})
}

}
\keyword{internal}
