% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_product_details.R
\name{get_product_details}
\alias{get_product_details}
\alias{get_product_url}
\alias{get_product_name}
\alias{get_product_name_unit}
\alias{get_product_unit}
\alias{get_product_image}
\title{Get Product Details}
\usage{
get_product_url(x)

get_product_name(x)

get_product_name_unit(x)

get_product_unit(x)

get_product_image(x)
}
\arguments{
\item{x}{Name or URL of a product.}
}
\description{
Get Product Details
}
\examples{
get_product_url("PLUS Aardbeien")

get_product_name("/product/plus-aardbeien-doos-225-g-244585")

get_product_name_unit("PLUS Aardbeien")

get_product_image("PLUS Aardbeien")
}
