% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_remote_functions.R
\name{plus_remote_functions}
\alias{plus_remote_functions}
\alias{plus_login}
\alias{plus_logout}
\alias{plus_add_products}
\alias{plus_open_browser}
\alias{plus_current_cart}
\alias{plus_checkout}
\alias{plus_current_url}
\alias{plus_page_reload}
\title{PLUS.nl Remote Functions}
\usage{
plus_login(
  credentials = getOption("plus_credentials", default =
    system.file("login_credentials.yaml", package = "shinyplus")),
  info = interactive()
)

plus_logout(info = interactive())

plus_add_products(x, quantity = 1, info = interactive(), ...)

plus_current_cart(..., info = interactive())

plus_checkout()

plus_current_url(b = plus_env$browser)

plus_open_browser()

plus_page_reload()
}
\arguments{
\item{credentials}{Path to a YAML file containing fields \code{email} and \code{password}, or a \link{list} contains those names. Can be set with \code{options(plus_credentials = "...")}}

\item{info}{Logical to print info, default is \code{TRUE} in interactive sessions.}

\item{x}{Name of the product, such as "PLUS Houdbare Halfvolle Melk Pak 1000 ml", or URL of the product, such as "plus-houdbare-halfvolle-melk-pak-1000-ml-957806".}

\item{quantity}{Number of items to add to cart.}

\item{...}{arguments passed to \code{\link[=plus_login]{plus_login()}}}
}
\description{
These functions allow remote access to the \href{https://www.plus.nl}{PLUS.nl website} through the \code{\link[chromote:Chromote]{Chromote}} class.
}
\examples{
\dontrun{
plus_login()

plus_add_products("Houdbare halfvolle melk pak 1000 ml")
}
}
