% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_remote_functions.R
\name{plus_remote_functions}
\alias{plus_remote_functions}
\alias{plus_login}
\alias{plus_logout}
\alias{plus_add_product}
\alias{plus_open_browser}
\title{PLUS.nl Remote Functions}
\usage{
plus_login(
  credentials_file = getOption("plus_credentials", default =
    system.file("login_credentials.yaml", package = "shinyplus")),
  info = interactive()
)

plus_logout(info = interactive())

plus_add_product(
  product_name = NULL,
  product_id = NULL,
  product_url = NULL,
  quantity = 1,
  info = interactive(),
  ...
)

plus_open_browser()
}
\arguments{
\item{credentials_file}{Path to a YAML file containing fields \code{email} and \code{password}. Can be set with \code{options(plus_credentials = "...")}}

\item{info}{Logical to print info, default is \code{TRUE} in interactive sessions.}

\item{product_name}{Name of the product, such as "PLUS Houdbare Halfvolle Melk Pak 1000 ml".}

\item{product_id}{PLUS ID of the product, such as "957806".}

\item{product_url}{PLUS URL of the product, such as "plus-houdbare-halfvolle-melk-pak-1000-ml-957806".}

\item{quantity}{Number of items to add to basket.}

\item{...}{arguments passed to \code{\link[=plus_login]{plus_login()}}}
}
\description{
These functions allow remote access to the \href{https://www.plus.nl}{PLUS.nl website} through the \code{\link[chromote:Chromote]{Chromote}} class.
}
\examples{
\dontrun{
plus_login()

plus_add_product(product_name = "PLUS Houdbare Halfvolle Melk Pak 1000 ml")
}
}
