% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_current_product_list_from_html.R
\name{update_current_product_list_from_html}
\alias{update_current_product_list_from_html}
\title{Update Product List}
\usage{
update_current_product_list_from_html(html_txt)
}
\arguments{
\item{html_txt}{Text to include in project list, as HTML element containing product entries as \verb{<a>} elements.}
}
\description{
Update the product list to be used in \code{\link[=shinyplus]{shinyplus()}}.
}
\details{
Steps to create local products list file:
\enumerate{
\item Go to:
\itemize{
\item \url{https://plus.nl/eerder-gekochte-producten} for personal, previously purchased products
\item \url{https://www.plus.nl/producten} for all products
\item \url{https://www.plus.nl/zoekresultaten?SearchTerm=YOURSEARCHTERM} to search for products
}
\item Scroll down as far as you like to load more products
\item Open the Web Info panel and copy a high-level HTML element to the clipboard (so that all \verb{<a>} elements of products are contained within)
\item Run:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{html_txt <- clipr::read_clip()
new_product_list <- update_current_product_list_from_html(html_txt)

# this is the local filepath for products (the R package does not contain product data)
path <- file.path(shinyplus:::plus_env$data_dir, "product_list.rds")

# make a backup
file.copy(from = path, to = paste0(path, ".", format(Sys.time(), "\%Y\%m\%d-\%H\%M\%S"), ".bak"))

# save new product list
saveRDS(new_product_list, path)
}\if{html}{\out{</div>}}
}

This process can be repeated as often as needed.
}
